requirements.txt
# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.13

# Vector Store
chromadb==0.4.22

# Ollama Integration
ollama==0.1.6

# PDF Processing
pypdf==4.0.1
pdfplumber==0.10.4

# Word Document Processing
python-docx==1.1.0

# Text Processing
tiktoken==0.5.2

# UI Framework
streamlit==1.30.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# Optional: For better PDF extraction
PyMuPDF==1.23.21
